Ancillary and smoothing coefficient improvements (#1359)
* Code changes

Remove the option to provide a coefficient argument. As the final result 
is rescaled between the min and max values the coefficient used makes no 
difference to the output.

Add the power used in the power law as a metadata attribute so that this 
is recorded.

* Moved ancillary creation from utilities to generate_ancillaries. Added an acceptance test for using a non-default power when generating smoothing coefficients.

* Review changes.

* Update checksum.
