+++ improver/generate_ancillaries/generate_orographic_smoothing_coefficients.py
@@ -28,16 +28,14 @@
 # CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
-"""A module for creating ancillary data"""
+"""A module for creating orographic smoothing coefficients"""
 
 import operator
-import warnings
 
 import iris
 import numpy as np
 
 from improver import BasePlugin
-from improver.constants import TRIPLE_PT_WATER
 from improver.metadata.constants.attributes import MANDATORY_ATTRIBUTE_DEFAULTS
 from improver.metadata.utilities import create_new_diagnostic_cube
 from improver.utilities.cube_manipulation import enforce_coordinate_ordering
@@ -58,11 +56,10 @@ class OrographicSmoothingCoefficients(BasePlugin):
     neighbouring cell.
 
     The smoothing coefficients are calculated from the orography gradient using
-    a simple equation and the user defined values for coefficient and power:
+    a simple equation with the user defined value for the power:
 
     .. math::
-        \\rm{smoothing\\_coefficient} = \\rm{coefficient} \\times
-        \\rm{gradient}^{\\rm{power}}
+        \\rm{smoothing\\_coefficient} = \\rm{gradient}^{\\rm{power}}
 
     The resulting values are scaled between min_gradient_smoothing_coefficient and
     max_gradient_smoothing_coefficient to give the desired range of
@@ -81,7 +78,6 @@ class OrographicSmoothingCoefficients(BasePlugin):
         self,
         min_gradient_smoothing_coefficient=0.5,
         max_gradient_smoothing_coefficient=0.0,
-        coefficient=1,
         power=1,
         use_mask_boundary=False,
         invert_mask=False,
@@ -100,10 +96,8 @@ class OrographicSmoothingCoefficients(BasePlugin):
                 where the orography gradient is a maximum. Generally this number
                 will be smaller than the min_gradient_smoothing_coefficient as
                 quantities are likely to be smoothed less across complex terrain.
-            coefficient (float):
-                The coefficient for the smoothing_coefficient equation
             power (float):
-                What power you want for your smoothing_coefficient equation
+                The power to be used in the smoothing_coefficient equation
             use_mask_boundary (bool):
                 A mask can be provided to this plugin to define a region in which
                 smoothing coefficients are set to zero, i.e. no smoothing. If this
@@ -137,7 +131,6 @@ class OrographicSmoothingCoefficients(BasePlugin):
 
         self.max_gradient_smoothing_coefficient = max_gradient_smoothing_coefficient
         self.min_gradient_smoothing_coefficient = min_gradient_smoothing_coefficient
-        self.coefficient = coefficient
         self.power = power
         self.use_mask_boundary = use_mask_boundary
         self.mask_comparison = operator.ge
@@ -182,9 +175,9 @@ class OrographicSmoothingCoefficients(BasePlugin):
     def unnormalised_smoothing_coefficients(self, gradient_cube):
         """
         This generates initial smoothing_coefficient values from gradients
-        using a generalised power law, whose parameters are set at
-        initialisation. Current defaults give an output
-        smoothing_coefficients_cube equal to the input gradient_cube.
+        using a simple power law, for which the power is set at initialisation.
+        Using a power of 1 gives an output smoothing_coefficients_cube with
+        values equal to the input gradient_cube.
 
         Args:
             gradient_cube (iris.cube.Cube):
@@ -194,10 +187,9 @@ class OrographicSmoothingCoefficients(BasePlugin):
             numpy.ndarray:
                 An array containing the unscaled smoothing_coefficients.
         """
-        return self.coefficient * np.abs(gradient_cube.data) ** self.power
+        return np.abs(gradient_cube.data) ** self.power
 
-    @staticmethod
-    def create_coefficient_cube(data, template, cube_name, attributes):
+    def create_coefficient_cube(self, data, template, cube_name, attributes):
         """
         Update metadata in smoothing_coefficients cube. Remove any time
         coordinates and rename.
@@ -224,6 +216,7 @@ class OrographicSmoothingCoefficients(BasePlugin):
 
         attributes["title"] = "Recursive filter smoothing coefficients"
         attributes.pop("history", None)
+        attributes["power"] = self.power
 
         return create_new_diagnostic_cube(
             cube_name,
@@ -360,143 +353,3 @@ class OrographicSmoothingCoefficients(BasePlugin):
             enforce_coordinate_ordering(smoothing_coefficient, original_order)
 
         return smoothing_coefficients
-
-
-class SaturatedVapourPressureTable(BasePlugin):
-
-    """
-    Plugin to create a saturated vapour pressure lookup table.
-    """
-
-    MAX_VALID_TEMPERATURE = 373.0
-    MIN_VALID_TEMPERATURE = 173.0
-
-    def __init__(self, t_min=183.15, t_max=338.25, t_increment=0.1):
-        """
-        Create a table of saturated vapour pressures that can be interpolated
-        through to obtain an SVP value for any temperature within the range
-        t_min --> (t_max - t_increment).
-
-        The default min/max values create a table that provides SVP values
-        covering the temperature range -90C to +65.1C. Note that the last
-        bin is not used, so the SVP value corresponding to +65C is the highest
-        that will be used.
-
-        Args:
-            t_min (float):
-                The minimum temperature for the range, in Kelvin.
-            t_max (float):
-                The maximum temperature for the range, in Kelvin.
-            t_increment (float):
-                The temperature increment at which to create values for the
-                saturated vapour pressure between t_min and t_max.
-        """
-        self.t_min = t_min
-        self.t_max = t_max
-        self.t_increment = t_increment
-
-    def __repr__(self):
-        """Represent the configured plugin instance as a string."""
-        result = (
-            "<SaturatedVapourPressureTable: t_min: {}; t_max: {}; "
-            "t_increment: {}>".format(self.t_min, self.t_max, self.t_increment)
-        )
-        return result
-
-    def saturation_vapour_pressure_goff_gratch(self, temperature):
-        """
-        Saturation Vapour pressure in a water vapour system calculated using
-        the Goff-Gratch Equation (WMO standard method).
-
-        Args:
-            temperature (numpy.ndarray):
-                Temperature values in Kelvin. Valid from 173K to 373K
-
-        Returns:
-            numpy.ndarray:
-                Corresponding values of saturation vapour pressure for a pure
-                water vapour system, in hPa.
-
-        References:
-            Numerical data and functional relationships in science and
-            technology. New series. Group V. Volume 4. Meteorology.
-            Subvolume b. Physical and chemical properties of the air, P35.
-        """
-        constants = {
-            1: 10.79574,
-            2: 5.028,
-            3: 1.50475e-4,
-            4: -8.2969,
-            5: 0.42873e-3,
-            6: 4.76955,
-            7: 0.78614,
-            8: -9.09685,
-            9: 3.56654,
-            10: 0.87682,
-            11: 0.78614,
-        }
-        triple_pt = TRIPLE_PT_WATER
-
-        # Values for which method is considered valid (see reference).
-        # WetBulbTemperature.check_range(temperature.data, 173., 373.)
-        if (
-            temperature.max() > self.MAX_VALID_TEMPERATURE
-            or temperature.min() < self.MIN_VALID_TEMPERATURE
-        ):
-            msg = "Temperatures out of SVP table range: min {}, max {}"
-            warnings.warn(msg.format(temperature.min(), temperature.max()))
-
-        svp = temperature.copy()
-        for cell in np.nditer(svp, op_flags=["readwrite"]):
-            if cell > triple_pt:
-                n0 = constants[1] * (1.0 - triple_pt / cell)
-                n1 = constants[2] * np.log10(cell / triple_pt)
-                n2 = constants[3] * (
-                    1.0 - np.power(10.0, (constants[4] * (cell / triple_pt - 1.0)))
-                )
-                n3 = constants[5] * (
-                    np.power(10.0, (constants[6] * (1.0 - triple_pt / cell))) - 1.0
-                )
-                log_es = n0 - n1 + n2 + n3 + constants[7]
-                cell[...] = np.power(10.0, log_es)
-            else:
-                n0 = constants[8] * ((triple_pt / cell) - 1.0)
-                n1 = constants[9] * np.log10(triple_pt / cell)
-                n2 = constants[10] * (1.0 - (cell / triple_pt))
-                log_es = n0 - n1 + n2 + constants[11]
-                cell[...] = np.power(10.0, log_es)
-
-        return svp
-
-    def process(self):
-        """
-        Create a lookup table of saturation vapour pressure in a pure water
-        vapour system for the range of required temperatures.
-
-        Returns:
-            iris.cube.Cube:
-               A cube of saturated vapour pressure values at temperature
-               points defined by t_min, t_max, and t_increment (defined above).
-        """
-        temperatures = np.arange(
-            self.t_min, self.t_max + 0.5 * self.t_increment, self.t_increment
-        )
-        svp_data = self.saturation_vapour_pressure_goff_gratch(temperatures)
-
-        temperature_coord = iris.coords.DimCoord(
-            temperatures, "air_temperature", units="K"
-        )
-
-        # Output of the Goff-Gratch is in hPa, but we want to return in Pa.
-        svp = iris.cube.Cube(
-            svp_data,
-            long_name="saturated_vapour_pressure",
-            units="hPa",
-            dim_coords_and_dims=[(temperature_coord, 0)],
-        )
-        svp.convert_units("Pa")
-        svp.attributes["minimum_temperature"] = self.t_min
-        svp.attributes["maximum_temperature"] = self.t_max
-        svp.attributes["temperature_increment"] = self.t_increment
-
-        return svp
